fix: circular deps in certification
